+++ src/menu/year.ts
@@ -84,7 +84,7 @@ export const footerIn: Menu = [
     },
     {
         name: 'Impressum',
-        url: '/impressum',
+        url: '/imprint',
         description: 'Klicke hier um zum impressum zu gelangen'
     },
 ];
@@ -102,7 +102,7 @@ export const footerOut: Menu = [
     },
     {
         name: 'Impressum',
-        url: '/impressum',
+        url: '/imprint',
         description: 'Klicke hier um zum impressum zu gelangen'
     },
 ];
